remove the page view from the home page; use an enum to manage the selected tab
